decouple pbtree.pagemanage to individual classes
